+++ src/main/java/de/markusbordihn/easymobfarm/loot/LootManager.java
@@ -40,6 +40,7 @@ import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
+import net.minecraft.world.item.enchantment.Enchantments;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.storage.loot.LootContext;
 import net.minecraft.world.level.storage.loot.LootTable;
@@ -55,6 +56,7 @@ import net.minecraftforge.registries.ForgeRegistries;
 import de.markusbordihn.easymobfarm.Constants;
 import de.markusbordihn.easymobfarm.config.CommonConfig;
 import de.markusbordihn.easymobfarm.config.mobs.BeeAnimal;
+import de.markusbordihn.easymobfarm.config.mobs.BossMonster;
 import de.markusbordihn.easymobfarm.config.mobs.HostileMonster;
 import de.markusbordihn.easymobfarm.config.mobs.HostileNetherMonster;
 import de.markusbordihn.easymobfarm.config.mobs.PassiveAnimal;
@@ -126,13 +128,21 @@ public class LootManager {
       LootContext.Builder lootBuilder = null;
       if (weaponItem != null && !weaponItem.isEmpty()) {
         // Kills with weapons has automatically a higher luck.
-        lootBuilder = new LootContext.Builder(serverLevel).withLuck(0.6F)
+        // Looting and Luck of See automatically increase mob loot.
+        float dropLuck = 0.7f;
+        if (weaponItem.getEnchantmentLevel(Enchantments.MOB_LOOTING) > 0) {
+          dropLuck = dropLuck + (weaponItem.getEnchantmentLevel(Enchantments.MOB_LOOTING) * 0.3f);
+        } else if (weaponItem.getEnchantmentLevel(Enchantments.FISHING_LUCK) > 0) {
+          dropLuck = dropLuck + (weaponItem.getEnchantmentLevel(Enchantments.FISHING_LUCK) * 0.3f);
+        }
+        lootBuilder = new LootContext.Builder(serverLevel).withLuck(dropLuck)
             .withParameter(LootContextParams.DAMAGE_SOURCE, DamageSource.playerAttack(player))
             .withParameter(LootContextParams.DIRECT_KILLER_ENTITY, player)
             .withParameter(LootContextParams.KILLER_ENTITY, player)
             .withParameter(LootContextParams.LAST_DAMAGE_PLAYER, player)
             .withParameter(LootContextParams.ORIGIN, player.position())
-            .withParameter(LootContextParams.THIS_ENTITY, player);
+            .withParameter(LootContextParams.THIS_ENTITY, player)
+            .withParameter(LootContextParams.TOOL, weaponItem);
       } else {
         lootBuilder = new LootContext.Builder(serverLevel).withLuck(0.5F)
             .withParameter(LootContextParams.DAMAGE_SOURCE, DamageSource.playerAttack(player))
@@ -240,6 +250,12 @@ public class LootManager {
             || (Constants.PRODUCTIVE_BEES_LOADED && BeeAnimal.ProductiveBees.contains(mobType)))
         && random.nextInt(3) == 0) {
       lootDrops.add(new ItemStack(Items.HONEYCOMB));
+    }
+
+    // Wither nether start drop support.
+    if (Boolean.TRUE.equals(COMMON.witherDropNetherStar.get()) && mobType.equals(BossMonster.WITHER)
+        && random.nextInt(9) == 0) {
+      lootDrops.add(new ItemStack(Items.NETHER_STAR));
     }
 
     // Check each single loot drop.